requirements.txt
h5py
hydra-core
numba
numpy
omegaconf
openpyxl
pandas
scipy
torch
tqdm
matplotlib
scikit-learn
seaborn